fixed f64 pos transition issues in other games

+++ exampleGames/dehexer/stateInjects.zig
@@ -6,7 +6,7 @@ pub var GRID_ID     : u32 = 0;
 pub var GRID_WIDTH  : i32 = 50;
 pub var GRID_HEIGHT : i32 = 25;
 
-pub var GRID_SCALE  : f32 = 50; // NOTE : Will be overwritten
+pub var GRID_SCALE  : f64 = 50; // NOTE : Will be overwritten
 
 pub const TILE_MINE_1 = def.e_tile_type.T1;
 pub const TILE_MINE_2 = def.e_tile_type.T2;
@@ -24,7 +24,7 @@ pub fn OnStart( ng : *def.Engine ) void
 pub fn OnOpen( ng : *def.Engine ) void
 {
   // Adjusting grid scalling to fit the screen
-  const scaleFactor : f32 = @floatFromInt( 1 + @max( GRID_WIDTH, GRID_HEIGHT * 2 ));
+  const scaleFactor : f64 = @floatFromInt( 1 + @max( GRID_WIDTH, GRID_HEIGHT * 2 ));
 
   const scale2 = def.getScreenSize().addVal( 128 ).divVal( scaleFactor ).?;
